Reorganize code into modules
