+++ hal/src/sercom/v2/uart/mod.rs
@@ -204,7 +204,7 @@ type Pads = uart::PadsFromIds<Sercom0, PA09, PA09>;
 //!
 //! Only the [`Uart`] struct can actually perform
 //! transactions. To do so, use the embedded HAL traits, like
-//! [`serial::Read`](Read) and [`serial::Write`](Write).
+//! [`serial::Read`] and [`serial::Write`].
 //!
 //! ```
 //! use nb::block;
@@ -339,6 +339,8 @@ type Pads = uart::PadsFromIds<Sercom0, PA09, PA09>;
 //! [`split`]: Uart::split
 //! [`join`]: Uart::join
 //! [`NoneT`]: crate::typelevel::NoneT
+//! [`serial::Write`]: embedded_hal::serial::Write
+//! [`serial::Read`]: embedded_hal::serial::Read
 #![cfg_attr(
     feature = "dma",
     doc = "
@@ -383,125 +385,33 @@ let (chan1, rx, rx_buffer) = rx_dma.wait();
 "
 )]
 
+#[cfg(any(feature = "samd11", feature = "samd21"))]
+#[path = "pads_thumbv6m.rs"]
+mod pads;
+
+#[cfg(feature = "min-samd51g")]
+#[path = "pads_thumbv7em.rs"]
+mod pads;
+
+pub use pads::*;
+
 mod reg;
 use reg::Registers;
 
-use super::*;
-use core::{
-    convert::{TryFrom, TryInto},
-    marker::PhantomData,
-};
+mod charsize;
+pub use charsize::*;
 
-use crate::{
-    target_device as pac,
-    time::Hertz,
-    typelevel::{Is, Sealed},
-};
-use bitflags::bitflags;
+mod flags;
+pub use flags::*;
 
-#[cfg(any(feature = "samd11", feature = "samd21"))]
-pub use crate::thumbv6m::sercom::v2::uart::*;
+mod config;
+pub use config::*;
 
-#[cfg(feature = "min-samd51g")]
-pub use crate::thumbv7em::sercom::v2::uart::*;
+pub mod impl_ehal;
 
-use embedded_hal::{
-    blocking,
-    serial::{Read, Write},
-};
-use nb::Error::WouldBlock;
-use num_traits::{AsPrimitive, PrimInt};
-
-//=============================================================================
-// Character size
-//=============================================================================
-
-/// Size of the SERCOM's `DATA` register
-#[cfg(any(feature = "samd11", feature = "samd21"))]
-pub type DataSize = u16;
-
-/// Size of the SERCOM's `DATA` register
-#[cfg(any(feature = "min-samd51g"))]
-pub type DataSize = u32;
-
-/// Type-level `enum` representing the UART character size
-///
-/// The UART character size affects the word size for the embedded HAL traits.
-/// Eight or less bit transactions use a `u8` word, while nine-bit transactions
-/// use a `u16` word.
-pub trait CharSize: Sealed {
-    /// Word size for the character size
-    type Word: 'static + PrimInt + AsPrimitive<DataSize>;
-
-    /// Bits to write into the `LENGTH` register
-    const BITS: u8;
-}
-
-/// Type-level `enum` indicating a [`CharSize`] that is not dynamic
-pub trait FixedCharSize: CharSize {}
-
-/// Type alias to recover the `Word` type from an implementation of [`CharSize`]
-pub type Word<C> = <C as CharSize>::Word;
-
-/// [`CharSize`] variant for 5-bit transactions
-pub enum FiveBit {}
-
-/// [`CharSize`] variant for 6-bit transactions
-pub enum SixBit {}
-
-/// [`CharSize`] variant for 7-bit transactions
-pub enum SevenBit {}
-
-/// [`CharSize`] variant for 8-bit transactions
-pub enum EightBit {}
-
-/// [`CharSize`] variant for 9-bit transactions
-pub enum NineBit {}
-
-/// Dynamic [`CharSize`] that can be changed on the fly
-pub enum DynCharSize {}
-
-impl Sealed for FiveBit {}
-impl CharSize for FiveBit {
-    type Word = u8;
-    const BITS: u8 = 0x5;
-}
-impl FixedCharSize for FiveBit {}
-
-impl Sealed for SixBit {}
-impl CharSize for SixBit {
-    type Word = u8;
-    const BITS: u8 = 0x6;
-}
-impl FixedCharSize for SixBit {}
-
-impl Sealed for SevenBit {}
-impl CharSize for SevenBit {
-    type Word = u8;
-    const BITS: u8 = 0x7;
-}
-impl FixedCharSize for SevenBit {}
-
-impl Sealed for EightBit {}
-impl CharSize for EightBit {
-    type Word = u8;
-    const BITS: u8 = 0x0;
-}
-impl FixedCharSize for EightBit {}
-
-impl Sealed for NineBit {}
-impl CharSize for NineBit {
-    type Word = u16;
-    const BITS: u8 = 0x1;
-}
-impl FixedCharSize for NineBit {}
-
-impl Sealed for DynCharSize {}
-impl CharSize for DynCharSize {
-    type Word = u16;
-    // Irrelevant for DynCharSize
-    const BITS: u8 = 0x0;
-}
+use crate::{sercom::v2::*, typelevel::Sealed};
+use core::{convert::TryInto, marker::PhantomData};
+use num_traits::AsPrimitive;
 
 //=============================================================================
 // Stop bits, parity, baud rate, bit order
@@ -634,129 +544,6 @@ impl BaudMode {
 }
 
 //=============================================================================
-// Interrupt flags
-//=============================================================================
-const DRE: u8 = 0x01;
-const TXC: u8 = 0x02;
-const RXC: u8 = 0x04;
-const RXS: u8 = 0x08;
-const CTSIC: u8 = 0x10;
-const RXBRK: u8 = 0x20;
-const ERROR: u8 = 0x80;
-
-/// Interrupt flags available for RX transactions
-pub const RX_FLAG_MASK: u8 = RXC | RXS | RXBRK | ERROR;
-/// Interrupt flags available for TX transactions
-pub const TX_FLAG_MASK: u8 = DRE | TXC;
-/// Interrupt flags available for Duplex transactions
-pub const DUPLEX_FLAG_MASK: u8 = RX_FLAG_MASK | TX_FLAG_MASK;
-
-bitflags! {
-    /// Interrupt bit flags for UART Rx transactions
-    ///
-    /// The available interrupt flags are `DRE`, `TXC`, `RXC`, `RXS`, `CTSIC`, `RXBRK` and
-    /// `ERROR`. The binary format of the underlying bits exactly matches the
-    /// INTFLAG bits.
-    pub struct Flags: u8 {
-        const DRE = DRE;
-        const TXC = TXC;
-        const RXC = RXC;
-        const RXS = RXS ;
-        const CTSIC = CTSIC;
-        const RXBRK = RXBRK;
-        const ERROR = ERROR;
-    }
-}
-
-//=============================================================================
-// Status flags
-//=============================================================================
-
-const PERR: u16 = 0x01;
-const FERR: u16 = 0x02;
-const BUFOVF: u16 = 0x04;
-const CTS: u16 = 0x08;
-const ISF: u16 = 0x10;
-const COLL: u16 = 0x20;
-
-/// Status flags available for RX transactions
-const RX_STATUS_MASK: u16 = PERR | FERR | BUFOVF | ISF | COLL;
-/// Status flags available for Duplex transactions
-const DUPLEX_STATUS_MASK: u16 = RX_STATUS_MASK;
-
-bitflags! {
-    /// Status flags for UART Rx transactions
-    ///
-    /// The available status flags are `PERR`, `FERR`, `BUFOVF`,
-    /// `CTS`, `ISF` and `COLL`.
-    /// The binary format of the underlying bits exactly matches
-    /// the STATUS bits.
-    pub struct Status: u16 {
-        const PERR = PERR;
-        const FERR = FERR;
-        const BUFOVF = BUFOVF;
-        const CTS = CTS;
-        const ISF = ISF;
-        const COLL = COLL;
-    }
-}
-
-//=============================================================================
-// Error
-//=============================================================================
-
-/// Errors available for UART transactions
-#[derive(Debug, Clone, Copy)]
-pub enum Error {
-    /// Detected a parity error
-    ParityError,
-    /// Detected a frame error
-    FrameError,
-    /// Detected a buffer overflow
-    Overflow,
-    /// Detected an inconsistent sync field
-    InconsistentSyncField,
-    /// Detected a collision
-    CollisionDetected,
-}
-
-impl TryFrom<Status> for () {
-    type Error = Error;
-
-    #[inline]
-    fn try_from(errors: Status) -> Result<(), Error> {
-        use Error::*;
-        if errors.contains(Status::PERR) {
-            Err(ParityError)
-        } else if errors.contains(Status::FERR) {
-            Err(FrameError)
-        } else if errors.contains(Status::BUFOVF) {
-            Err(Overflow)
-        } else if errors.contains(Status::ISF) {
-            Err(InconsistentSyncField)
-        } else if errors.contains(Status::COLL) {
-            Err(CollisionDetected)
-        } else {
-            Ok(())
-        }
-    }
-}
-
-impl From<Error> for Status {
-    #[inline]
-    fn from(err: Error) -> Self {
-        use Error::*;
-        match err {
-            ParityError => Status::PERR,
-            FrameError => Status::FERR,
-            Overflow => Status::BUFOVF,
-            InconsistentSyncField => Status::ISF,
-            CollisionDetected => Status::COLL,
-        }
-    }
-}
-
-//=============================================================================
 // Capability
 //=============================================================================
 
@@ -843,466 +630,6 @@ impl Capability for TxDuplex {
 impl Transmit for TxDuplex {}
 
 //=============================================================================
-// Config
-//=============================================================================
-
-/// A configurable, disabled UART peripheral
-///
-/// This `struct` represents a configurable UART peripheral in its disabled
-/// state. It is generic over the set of [`Pads`] and [`CharSize`].
-/// Upon creation, the [`Config`] takes ownership of the
-/// [`Sercom`] and resets it, returning it configured as an UART peripheral
-/// with a default configuration:
-///
-/// * [`EightBit`]
-/// * No parity
-/// * One stop bit
-/// * LSB-first
-///
-/// [`Config`] uses a builder-pattern API to configure the peripheral,
-/// culminating in a call to [`enable`], which consumes the [`Config`] and
-/// returns enabled [`Uart`]. The [`enable`] method is
-/// restricted to [`ValidConfig`]s.
-///
-/// [`enable`]: Config::enable
-pub struct Config<P, C = EightBit>
-where
-    P: ValidPads,
-    C: CharSize,
-{
-    registers: Registers<P::Sercom>,
-    pads: P,
-    chsize: PhantomData<C>,
-    freq: Hertz,
-}
-
-#[cfg(any(feature = "samd11", feature = "samd21"))]
-pub type Clock = pac::PM;
-
-#[cfg(feature = "min-samd51g")]
-pub type Clock = pac::MCLK;
-
-impl<P: ValidPads> Config<P> {
-    /// Create a new [`Config`] in the default configuration
-    ///
-    /// This function will enable the corresponding APB clock, reset the
-    /// [`Sercom`] peripheral, and return a [`Config`] in the default
-    /// configuration:
-    ///
-    /// * [`EightBit`] [`CharSize`]
-    /// * No parity
-    /// * One stop bit
-    /// * LSB-first
-    ///
-    /// [`Config`] takes ownership of the [`Sercom`] and [`Pads`].
-    ///
-    /// Users must configure GCLK manually. The `freq` parameter represents the
-    /// GCLK frequency for this [`Sercom`] instance.
-    #[inline]
-    pub fn new(clk: &Clock, mut sercom: P::Sercom, pads: P, freq: impl Into<Hertz>) -> Self {
-        sercom.enable_apb_clock(clk);
-        Self::default(sercom, pads, freq).bit_order(BitOrder::LsbFirst)
-    }
-
-    /// Create a new [`Config`] in the default configuration
-    #[inline]
-    fn default(sercom: P::Sercom, pads: P, freq: impl Into<Hertz>) -> Self {
-        let mut registers = Registers::new(sercom);
-        registers.swrst();
-
-        // Enable internal clock mode
-        registers.configure_mode();
-        registers.configure_pads(P::RXPO as u8, P::TXPO as u8);
-        registers.configure_charsize(EightBit::BITS);
-
-        Self {
-            registers,
-            pads,
-            chsize: PhantomData,
-            freq: freq.into(),
-        }
-    }
-}
-
-impl<P, C> Config<P, C>
-where
-    P: ValidPads,
-    C: CharSize,
-{
-    /// Change the [`Config`] [`CharSize`]
-    #[inline]
-    fn change<C2>(self) -> Config<P, C2>
-    where
-        C2: CharSize,
-    {
-        Config {
-            registers: self.registers,
-            pads: self.pads,
-            chsize: PhantomData,
-            freq: self.freq,
-        }
-    }
-
-    /// Trigger the [`Sercom`]'s SWRST and return a [`Config`] in the
-    /// default configuration.
-    #[inline]
-    pub fn reset(self) -> Config<P> {
-        Config::default(self.registers.free(), self.pads, self.freq)
-    }
-
-    /// Consume the [`Config`], reset the peripheral, and return the [`Sercom`]
-    /// and [`Pads`]
-    #[inline]
-    pub fn free(mut self) -> (P::Sercom, P) {
-        self.registers.swrst();
-        (self.registers.free(), self.pads)
-    }
-
-    /// Change the [`CharSize`].
-    #[inline]
-    pub fn char_size<C2: FixedCharSize>(mut self) -> Config<P, C2> {
-        self.registers.configure_charsize(C2::BITS);
-        self.change()
-    }
-
-    /// Change the [`CharSize`] to [`DynCharSize`]. The UART's character
-    /// size will be changed to the provided `C2`, which is a [`FixedCharSize`]
-    /// type, and can then be changed dynamically on an enabled [`Uart`]
-    /// without changing the underlying [`Config`]'s type through the
-    /// [`reconfigure`](Uart::reconfigure) method.
-    #[inline]
-    pub fn dyn_char_size<C2: FixedCharSize>(mut self) -> Config<P, DynCharSize> {
-        self.registers.configure_charsize(C2::BITS);
-        self.change()
-    }
-
-    /// Change the bit order of transmission (builder pattern version)
-    #[inline]
-    pub fn bit_order(mut self, bit_order: BitOrder) -> Self {
-        self.set_bit_order(bit_order);
-        self
-    }
-
-    /// Change the bit order of transmission (setter version)
-    #[inline]
-    pub fn set_bit_order(&mut self, bit_order: BitOrder) {
-        self.registers.set_bit_order(bit_order);
-    }
-
-    /// Get the current bit order
-    #[inline]
-    pub fn get_bit_order(&self) -> BitOrder {
-        self.registers.get_bit_order()
-    }
-
-    /// Change the parity setting (builder pattern version)
-    #[inline]
-    pub fn parity(mut self, parity: Parity) -> Self {
-        self.set_parity(parity);
-        self
-    }
-
-    /// Change the parity setting (setter version)
-    #[inline]
-    pub fn set_parity(&mut self, parity: Parity) {
-        self.registers.set_parity(parity);
-    }
-
-    /// Get the current parity setting
-    #[inline]
-    pub fn get_parity(&self) -> Parity {
-        self.registers.get_parity()
-    }
-
-    /// Change the stop bit setting (builder pattern version)
-    #[inline]
-    pub fn stop_bits(mut self, stop_bits: StopBits) -> Self {
-        self.set_stop_bits(stop_bits);
-        self
-    }
-
-    /// Change the stop bit setting (setter version)
-    #[inline]
-    pub fn set_stop_bits(&mut self, stop_bits: StopBits) {
-        self.registers.set_stop_bits(stop_bits);
-    }
-
-    /// Get the current stop bit setting
-    #[inline]
-    pub fn get_stop_bits(&self) -> StopBits {
-        self.registers.get_stop_bits()
-    }
-
-    /// Enable or disable the start of frame detector (builder pattern version)
-    ///
-    /// When set, the UART will generate interrupts for
-    /// RXC and/or RXS if these interrupt flags have been enabled.
-    #[inline]
-    pub fn start_of_frame_detection(mut self, enabled: bool) -> Self {
-        self.set_start_of_frame_detection(enabled);
-        self
-    }
-
-    /// Enable or disable the start of frame detector (setter version)
-    ///
-    /// When set, the UART will generate interrupts for
-    /// RXC and/or RXS if these interrupt flags have been enabled.
-    #[inline]
-    pub fn set_start_of_frame_detection(&mut self, enabled: bool) {
-        self.registers.set_start_of_frame_detection(enabled);
-    }
-
-    /// Get the current SOF detector setting
-    #[inline]
-    pub fn get_start_of_frame_detection(&self) -> bool {
-        self.registers.get_start_of_frame_detection()
-    }
-
-    /// Enable or disable the collision detector (builder pattern version)
-    ///
-    /// When set, the UART will detect collisions and update the
-    /// corresponding flag in the STATUS register.
-    #[inline]
-    pub fn collision_detection(mut self, enabled: bool) -> Self {
-        self.set_collision_detection(enabled);
-        self
-    }
-
-    /// Enable or disable the collision detector (setter version)
-    ///
-    /// When set, the UART will detect collisions and update the
-    /// corresponding flag in the STATUS register.
-    #[inline]
-    pub fn set_collision_detection(&mut self, enabled: bool) {
-        self.registers.set_collision_detection(enabled);
-    }
-
-    /// Get the current collision detector setting
-    #[inline]
-    pub fn get_collision_detection(&self) -> bool {
-        self.registers.get_collision_detection()
-    }
-
-    /// Set the baud rate (builder pattern version)
-    ///
-    /// This function will calculate the best BAUD register setting based on the
-    /// stored GCLK frequency and desired baud rate. The maximum baud rate is
-    /// GCLK frequency/oversampling. Values outside this range will saturate at
-    /// the maximum supported baud rate.
-    ///
-    /// Note that 3x oversampling is not supported.
-    #[inline]
-    pub fn baud<B: Into<Hertz>>(mut self, baud: B, mode: BaudMode) -> Self {
-        self.set_baud(baud, mode);
-        self
-    }
-
-    /// Set the baud rate (setter version)
-    ///
-    /// This function will calculate the best BAUD register setting based on the
-    /// stored GCLK frequency and desired baud rate. The maximum baud rate is
-    /// GCLK frequency/oversampling. Values outside this range will saturate at
-    /// the maximum supported baud rate.
-    ///
-    /// Note that 3x oversampling is not supported.
-    #[inline]
-    pub fn set_baud<B: Into<Hertz>>(&mut self, baud: B, mode: BaudMode) {
-        self.registers.set_baud(self.freq, baud, mode);
-    }
-
-    /// Get the contents of the `BAUD` register and the current baud mode. Note
-    /// that only the CONTENTS of `BAUD` are returned, and not the actual baud
-    /// rate. Refer to the datasheet to convert the `BAUD` register contents
-    /// into a baud rate.
-    #[inline]
-    pub fn get_baud(&self) -> (u16, BaudMode) {
-        self.registers.get_baud()
-    }
-
-    /// Control the buffer overflow notification (builder pattern version)
-    ///
-    /// If set to true, an [`Error::Overflow`] will be issued as soon as an
-    /// overflow occurs. Otherwise, it will not be issued until its place within
-    /// the data stream.
-    #[inline]
-    pub fn immediate_overflow_notification(mut self, set: bool) -> Self {
-        self.set_immediate_overflow_notification(set);
-        self
-    }
-
-    /// Control the buffer overflow notification (setter version)
-    ///
-    /// If set to true, an [`Error::Overflow`] will be issued as soon as an
-    /// overflow occurs. Otherwise, it will not be issued until its place within
-    /// the data stream.
-    #[inline]
-    pub fn set_immediate_overflow_notification(&mut self, set: bool) {
-        self.registers.set_immediate_overflow_notification(set);
-    }
-
-    /// Get the current immediate overflow notification setting
-    #[inline]
-    pub fn get_immediate_overflow_notification(&self) -> bool {
-        self.registers.get_immediate_overflow_notification()
-    }
-
-    /// Run in standby mode (builder pattern version)
-    ///
-    /// When set, the UART peripheral will run in standby mode. See the
-    /// datasheet for more details.
-    #[inline]
-    pub fn run_in_standby(mut self, set: bool) -> Self {
-        self.set_run_in_standby(set);
-        self
-    }
-
-    /// Run in standby mode (setter version)
-    ///
-    /// When set, the UART peripheral will run in standby mode. See the
-    /// datasheet for more details.
-    #[inline]
-    pub fn set_run_in_standby(&mut self, set: bool) {
-        self.registers.set_run_in_standby(set);
-    }
-
-    /// Get the current run in standby mode
-    #[inline]
-    pub fn get_run_in_standby(&self) -> bool {
-        self.registers.get_run_in_standby()
-    }
-
-    /// Enable or disable IrDA encoding (builder pattern version)
-    ///
-    /// The pulse length controls the minimum pulse length that is required for
-    /// a pulse to be accepted by the IrDA receiver with regards to the
-    /// serial engine clock period. See datasheet for more information.
-    #[inline]
-    pub fn irda_encoding(mut self, pulse_length: Option<u8>) -> Self {
-        self.set_irda_encoding(pulse_length);
-        self
-    }
-
-    /// Enable or disable IrDA encoding (setter version)
-    ///
-    /// The pulse length controls the minimum pulse length that is required for
-    /// a pulse to be accepted by the IrDA receiver with regards to the
-    /// serial engine clock period. See datasheet for more information.
-    #[inline]
-    pub fn set_irda_encoding(&mut self, pulse_length: Option<u8>) {
-        self.registers.set_irda_encoding(pulse_length);
-    }
-
-    /// Get the current IrDA encoding setting. The return type is the pulse
-    /// length wrapped in an [`Option`].
-    #[inline]
-    pub fn get_irda_encoding(&self) -> Option<u8> {
-        self.registers.get_irda_encoding()
-    }
-}
-
-impl<P: ValidPads> Config<P, DynCharSize> {
-    /// Dynamically change the character size
-    ///
-    /// The UART's character size will be changed to the provided `C2`, which is
-    /// a [`FixedCharSize`] type, without changing the type of the
-    /// underlying [`Config`].
-    #[inline]
-    pub fn set_dyn_char_size<C2: FixedCharSize>(&mut self) {
-        self.registers.configure_charsize(C2::BITS);
-    }
-}
-
-impl<P, C> Config<P, C>
-where
-    P: ValidPads,
-    C: CharSize,
-    Self: ValidConfig,
-{
-    /// Enable the UART peripheral and return a [`Uart`] struct.
-    ///
-    /// UART transactions are not possible until the peripheral is enabled.
-    /// This method is limited to [`ValidConfig`]s
-    #[inline]
-    pub fn enable(mut self) -> Uart<Self, P::Capability> {
-        self.registers.enable();
-        Uart {
-            config: self,
-            capability: PhantomData,
-        }
-    }
-}
-
-//=============================================================================
-// AnyConfig
-//=============================================================================
-
-/// Type class for all possible [`Config`] types
-///
-/// This trait uses the [`AnyKind`] trait pattern to create a [type class] for
-/// [`Config`] types. See the `AnyKind` documentation for more details on the
-/// pattern.
-///
-/// In addition to the normal, `AnyKind` associated types. This trait also
-/// copies the [`Sercom`] and [`Word`] types, to make it easier to apply
-/// bounds to these types at the next level of abstraction.
-///
-/// [`AnyKind`]: crate::typelevel#anykind-trait-patter
-pub trait AnyConfig: Sealed + Is<Type = SpecificConfig<Self>> {
-    type Sercom: Sercom;
-    type Pads: ValidPads<Sercom = Self::Sercom>;
-    type Word: 'static + PrimInt + AsPrimitive<DataSize>;
-    type CharSize: CharSize<Word = Self::Word>;
-}
-
-/// Type alias to recover the specific [`Config`] type from an implementation of
-/// [`AnyConfig`]
-pub type SpecificConfig<C> = Config<<C as AnyConfig>::Pads, <C as AnyConfig>::CharSize>;
-
-/// Type alias to recover the specific [`Sercom`] type from an implementation of
-/// [`AnyConfig`]
-pub type ConfigSercom<C> = <C as AnyConfig>::Sercom;
-
-impl<P, C> AsRef<Self> for Config<P, C>
-where
-    P: ValidPads,
-    C: CharSize,
-{
-    #[inline]
-    fn as_ref(&self) -> &Self {
-        self
-    }
-}
-
-impl<P, C> AsMut<Self> for Config<P, C>
-where
-    P: ValidPads,
-    C: CharSize,
-{
-    #[inline]
-    fn as_mut(&mut self) -> &mut Self {
-        self
-    }
-}
-
-impl<P, C> Sealed for Config<P, C>
-where
-    P: ValidPads,
-    C: CharSize,
-{
-}
-
-impl<P, C> AnyConfig for Config<P, C>
-where
-    P: ValidPads,
-    C: CharSize,
-{
-    type Sercom = P::Sercom;
-    type Word = C::Word;
-    type Pads = P;
-    type CharSize = C;
-}
-
-//=============================================================================
 // Uart
 //=============================================================================
 
@@ -1374,7 +701,8 @@ where
     /// * Since [`Duplex`] [`Uart`]s are [`Receive`] + [`Transmit`] they have
     ///   all flags available.
     ///
-    /// **Warning:** The implementation of of [`Write::flush`] waits on and
+    /// **Warning:** The implementation of of
+    /// [`Write::flush`](embedded_hal::serial::Write::flush) waits on and
     /// clears the `TXC` flag. Manually clearing this flag could cause it to
     /// hang indefinitely.
     #[inline]
@@ -1674,68 +1002,4 @@ where
     pub unsafe fn write_data(&mut self, data: DataSize) {
         self.config.as_mut().registers.write_data(data);
     }
-}
-
-//=============================================================================
-// Embedded HAL traits
-//=============================================================================
-
-impl<C, D> Read<C::Word> for Uart<C, D>
-where
-    C: ValidConfig,
-    D: Receive,
-    DataSize: AsPrimitive<C::Word>,
-{
-    type Error = Error;
-
-    /// Wait for an `RXC` flag, then read the word
-    #[inline]
-    fn read(&mut self) -> nb::Result<C::Word, Error> {
-        let flags = self.read_flags_errors()?;
-        if flags.contains(Flags::RXC) {
-            unsafe { Ok(self.read_data().as_()) }
-        } else {
-            Err(WouldBlock)
-        }
-    }
-}
-
-impl<C, D> Write<C::Word> for Uart<C, D>
-where
-    C: ValidConfig,
-    D: Transmit,
-{
-    type Error = core::convert::Infallible;
-
-    /// Wait for a `DRE` flag, then write a word
-    #[inline]
-    fn write(&mut self, word: C::Word) -> nb::Result<(), Self::Error> {
-        // Ignore buffer overflow errors
-        if self.read_flags().contains(Flags::DRE) {
-            unsafe { self.write_data(word.as_()) };
-            Ok(())
-        } else {
-            Err(WouldBlock)
-        }
-    }
-
-    /// Wait for a `TXC` flag
-    #[inline]
-    fn flush(&mut self) -> nb::Result<(), Self::Error> {
-        // Ignore buffer overflow errors
-        if self.read_flags().contains(Flags::TXC) {
-            self.clear_flags(Flags::TXC);
-            Ok(())
-        } else {
-            Err(WouldBlock)
-        }
-    }
-}
-
-impl<C, D> blocking::serial::write::Default<C::Word> for Uart<C, D>
-where
-    C: ValidConfig,
-    D: Transmit,
-    Uart<C, D>: Write<C::Word>,
-{
 }
