+++ hal/src/thumbv6m/sercom/v2.rs
@@ -1,3 +1,2 @@
 pub mod impl_pad;
 pub mod spi;
-pub(crate) mod uart;
